Add files via upload
Analysis of Kickstarter Challenge

Section I: Overview of Project - The Purpose of this Analysis

Introduction

The purpose of the analysis provided within this workbook is to inform and guide our client, Louise, as she navigates the process of producing her play “Fever”.  At the onset of our analysis process, Louise informed us that her fundraising efforts came close to her goal in what she describes as “a short amount of time.”  The purpose of our continued efforts is to contrast how her fundraising efforts stack up to similar campaigns.

Specific Parameters

Louise had accumulated data from fundraising campaigns for projects of many different varieties, countries, and production mediums.  Although it was a blessing to have such a vast array of data to work with, this amalgamation of information did not adequately serve the specific purpose of informing Louise about projects that would be comparable to hers.  Thus, the formulas and data sets constructed for this project reflect plays that would attract a similar market to the one Louise has produced.

Section II:Analysis and Challenges

Theater Outcomes Based on Launch Date

The first things Louise wanted to know about was the most frequently successful month of the year to launch a fundraising campaign.  To analyze this, we created a Pivot Table incorporating the Parent Category, Years, and Outcomes columns of the “Kickstarter” sheet.  Using the filters, we displayed the “Count of Outcomes” with the row labels being the months of the year.  From there, the columns were broken into “Successful”, “Failed”, “Canceled”, and “Grand Total.”













This analysis then generated the data that was used to create this line graph:



After we completed the analysis of “Theater Outcomes by Launch Date” we shifted our attention to the specific goals these fundraising initiatives set.

Using the (CountIfs) function in the “Kickstarter” data set, we evaluated the number of “successful”, “failed”, and “canceled” projects based on the range of their fundraising goals.  We picked 12 different ranges from “Less Than 1000” to “50000 or More” to evaluate how much of an impact the goal range had on the success or failure rate of these different fundraising efforts.  Once we were able to get a count for each category, we use the (Sum) function for the total number of projects in that goal range.  From there, we used a simple formula to divide each column by the total to turn it into a percentage of the total projects for that range.  From the “Percentage Successful”, “Percentage Failed”, and “Percentage Canceled” columns, we were able to generate this chart:





Section III: Results

What are two conclusions you can draw about the Theater Outcomes by Launch Date?

The first conclusion that can confidently be drawn regarding the outcomes of theatrical productions based on their launch date is that May is the month with the most volatility for fundraising campaigns.  We know this because looking at the line chart we see that May has both the highest percentage of “successful” launches as well as the highest percentage of “failed” launches.

The second conclusion we can glean from the data is that after the month of May, the rate of success for launches goes down drastically month by month to the end of the year with the only exception being the month of October.  Following the “successful” data line found in blue, it is clear that as each month passes, a launch will be less likely to succeed.

What can you conclude about the Outcomes based on Goals?

The “Outcomes” lines do not follow as consistent of a pattern as the launch data.  However, we know with certainty that a goal is most likely to be met if it is below $1,000.  It is also most likely to fail if it falls within the $45,000 to $44,999 range.
What are some limitation of this dataset?

First, we are limited by the fact we do not have any specific data as to the mitigating circumstances around the fundraising campaigns for these plays.  We do not know about the specific regions of the country they took place in, how much was invested in effective marketing strategies, and the actual methods of approach each of the fundraisers took in their efforts.  All of these things would be elements that would impact Louise’s approach to fundraising.

We also do not have data about the specific regions the data is pulled from for the outcomes by launch date.  A play might do well in summer in a tourist location, but the same play with the same fundraising manager might fail in poor weather.  These sorts of outside influences that have a big impact on data gathering are just not available to us to build a background to our data analysis.

What are some other possible tables and/or graphs that we could create?

The data set that we did not explore at all that could be useful to Louise would be the “Pledged” column.  If we were able to show Louise a breakdown of what she would need in initial investment money to be pledged by outside donors, we could more successfully guide her plans for how much she would ultimately need to raise.